finished moving things
