+++ src/DemoPage/DemoMainContent/DemoTranslateText/components/DemoSentence/DemoSentence.css
